feat: Add script to upload batch of files to Azure Storage
Added a new script `az-upload-batch.sh` to upload a batch of files to Azure Storage. This script uses the `az storage blob upload-batch` command to upload all files in the `./locales` directory to the `acf-i18n` container in the `` storage account. The `--auth-mode login` flag is used for authentication and the `--overwrite` flag is used to overwrite any existing files with the same name.